+++ PliP/src/com/plip/persistence/dao/impls/TrayDaoImpl.java
@@ -7,6 +7,7 @@ import org.hibernate.SessionFactory;
 import org.hibernate.Transaction;
 
 import com.plip.persistence.dao.interfaces.TrayDao;
+import com.plip.persistence.exceptions.TrayNotFoundException;
 import com.plip.persistence.managers.DaoManager;
 import com.plip.persistence.model.Position;
 import com.plip.persistence.model.Status;
@@ -39,7 +40,7 @@ public class TrayDaoImpl implements TrayDao {
 	}
 
 	@Override
-	public Tray getTray(long idTray) {
+	public Tray getTray(long idTray) throws TrayNotFoundException {
 		SessionFactory factory = DaoManager.createSessionFactory();
 		Session session = factory.openSession();
 		Transaction tx = null;
@@ -49,7 +50,11 @@ public class TrayDaoImpl implements TrayDao {
 			Query query = session
 					.createQuery("FROM Tray where idTray = :id");
 			query.setParameter("id", idTray);
+			if(query.list().size()!=0){
 			tray = (Tray) query.list().get(0);
+			}else{
+				throw new TrayNotFoundException();
+			}
 			tx.commit();
 		} catch (HibernateException e) {
 			if (tx != null)
@@ -62,7 +67,7 @@ public class TrayDaoImpl implements TrayDao {
 	}
 
 	@Override
-	public void updateTray(Tray tray) {
+	public void updateTray(Tray tray) throws TrayNotFoundException {
 		SessionFactory factory = DaoManager.createSessionFactory();
 		Session session = factory.openSession();
 		Transaction tx = null;
@@ -70,9 +75,13 @@ public class TrayDaoImpl implements TrayDao {
 			tx = session.beginTransaction();
 			Tray tr = (Tray) session.get(Tray.class,
 					tray.getIdTray());
+			if(tr!=null){
 			tr.setCode(tray.getCode());
 			tr.setPage(tray.getPage());
 			session.update(tr);
+			}else{
+				throw new TrayNotFoundException();
+			}
 			tx.commit();
 		} catch (HibernateException e) {
 			if (tx != null)
@@ -91,7 +100,9 @@ public class TrayDaoImpl implements TrayDao {
 		try {
 			tx = session.beginTransaction();
 			Tray tray = (Tray) session.get(Tray.class, trayId);
+			if(tray!=null){
 			session.delete(tray);
+			}
 			tx.commit();
 		} catch (HibernateException e) {
 			if (tx != null)
